Fix format warning

+++ src/fast-aleck/fast-aleck-main.c
@@ -91,7 +91,8 @@ int main(int argc, char **argv)
 	gettimeofday(&tp2, NULL);
 
 	// Write
-	fprintf(stderr, "done in %li.%06is\n", tp2.tv_sec - tp1.tv_sec, tp2.tv_usec - tp1.tv_usec);
+	fprintf(stderr, "done in %li.%06is\n",
+        (int) (tp2.tv_sec - tp1.tv_sec, tp2.tv_usec - tp1.tv_usec));
 	fputs(out, stdout);
 
 	return 0;
